requirements.txt
streamlit
pyttsx3
PyPDF2
sounddevice
scipy
numpy
openai-whisper
torch
ollama
